requirements-server.txt
langchain_openai
openai==1.11.1
unstructured==0.11.8
trafilatura==1.8.0
requests
fastapi
newsapi-python
uvicorn